+++ src/cmd_fsn_voxel.rs
@@ -79,7 +79,7 @@ fn build_voxel(
     let max_dimension = dimensions.x.max(dimensions.y).max(dimensions.z);
 
     let radius = max_dimension * radius_multiplier; // unscaled
-    let scale = (divisions / max_dimension) as f32;
+    let scale = divisions / max_dimension;
 
     if verbose {
         println!(
@@ -92,7 +92,7 @@ fn build_voxel(
             divisions,
             max_dimension,
             scale,
-            (max_dimension as f32) * scale
+            max_dimension * scale
         );
         println!();
     }
