+++ src/cmd_voronoi.rs
@@ -345,7 +345,7 @@ fn parse_input(
 > {
     let mut aabb = linestring::linestring_3d::Aabb3::<f64>::default();
     for v in input_pb_model.vertices.iter() {
-        aabb.update_point(cgmath::Point3::new(v.x as f64, v.y as f64, v.z as f64))
+        aabb.update_point(cgmath::Point3::new(v.x, v.y, v.z))
     }
 
     let (plane, transform, vor_aabb)= centerline::get_transform_relaxed(
